In this commit, the Sign-up function is implemented correctly.
First, don't forget to add the <router-outlet> tag at the app component file, otherwise the correct page won't load even you have the correct router setting and urls.

Second, at the signup component, we made some validation for each field of the register box. It will display the message if the input are not validated, and it even goes red, setting on styles.css. This register box uses FormGroup class to create.

Since we want to send the obj to our Spring backend, we warp the data into a signup-required.payload.ts, it contains three fields that input from register box. Then we use HttpClient to send this obj to our Spring backend.