In this commit, we implement the token passing and refreshing here, as well as the post pages
The token-interceptor.ts is in charge of dealing with all kinds of token processing.

Basically it's like a java servlet filter. Give https request necessary header with jwt token.

It also refresh the jwt token use refresh token necessary.

The Post page is working right now, after logged in. Some issues with CROS security on chrome, tested with a chrome that disable the CROS.